+++ package.json
@@ -7,9 +7,10 @@
     "a0deploy": "lib/index.js"
   },
   "scripts": {
+    "aa": "",
     "lint:js": "eslint --ignore-path .gitignore --ignore-pattern webpack .",
-    "test": "npm run test:pre && cross-env NODE_ENV=test nyc mocha test/mocha.js ./test/**/*.tests.js ./test/*.tests.js",
-    "test:watch": "cross-env NODE_ENV=test mocha test/mocha.js ./test/**/*.tests.js ./test/*.tests.js --watch",
+    "test": "npm run test:pre && cross-env NODE_ENV=test nyc mocha --recursive --require babel-register test",
+    "test:watch": "cross-env NODE_ENV=test mocha --recursive --require babel-register test --watch",
     "test:pre": "npm run test:clean && npm run lint:js",
     "test:clean": "rimraf ./coverage && rimraf ./.nyc_output",
     "build": "babel src -d lib",
@@ -49,21 +50,21 @@
     "yargs": "^12.0.1"
   },
   "devDependencies": {
-    "chai": "^3.5.0",
-    "chai-as-promised": "^5.3.0",
-    "cross-env": "^3.1.4",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-import-resolver-babel-module": "^4.0.0",
-    "eslint-import-resolver-webpack": "^0.10.0",
-    "eslint-plugin-import": "^2.12.0",
     "babel-eslint": "^8.2.3",
     "babel-plugin-dynamic-import-node": "^1.2.0",
     "babel-plugin-module-resolver": "^3.1.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.5",
+    "chai": "^4.1.2",
+    "chai-as-promised": "^7.1.1",
+    "cross-env": "^3.1.4",
     "eslint": "^4.19.1",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-import-resolver-babel-module": "^4.0.0",
+    "eslint-import-resolver-webpack": "^0.10.0",
+    "eslint-plugin-import": "^2.12.0",
     "expect": "^1.20.2",
     "mkdirp": "^0.5.1",
-    "mocha": "^3.0.2",
+    "mocha": "^5.2.0",
     "nyc": "^10.0.0",
     "rmdir-sync": "^1.0.1",
     "supertest": "^2.0.1"
