+++ src/context/directory/handlers/resourceServers.js
@@ -1,10 +1,11 @@
 import path from 'path';
 import { unifyScripts, constants } from 'auth0-source-control-extension-tools';
 
-import { groupFiles, parseFileGroup } from 'src/utils';
+import { existsMustBeDir, groupFiles, parseFileGroup } from 'src/utils';
 
 export default function parse(folder, mappings) {
   const resourceFolder = path.join(folder, constants.RESOURCE_SERVERS_DIRECTORY);
+  existsMustBeDir(resourceFolder);
   const filesGrouped = groupFiles(resourceFolder);
 
   const resources = Object.entries(filesGrouped)
