+++ src/context/directory/handlers/rules.js
@@ -2,7 +2,7 @@ import path from 'path';
 import { unifyScripts, constants } from 'auth0-source-control-extension-tools';
 
 import { logger } from 'src/logger';
-import { loadFile, groupFiles } from 'src/utils';
+import { loadFile, groupFiles, existsMustBeDir } from 'src/utils';
 
 
 function parseFileGroup(name, files, mappings) {
@@ -12,8 +12,8 @@ function parseFileGroup(name, files, mappings) {
     const content = loadFile(file, mappings);
     const { ext } = path.parse(file);
     if (ext === '.json') {
-      rule.metadataFile = true;
-      rule.metadata = content;
+      rule.metadataFile = content;
+      rule.metadata = true;
     } else if (ext === '.js') {
       rule.script = true;
       rule.scriptFile = content;
@@ -27,6 +27,7 @@ function parseFileGroup(name, files, mappings) {
 
 export default function parse(folder, mappings) {
   const rulesFolder = path.join(folder, constants.RULES_DIRECTORY);
+  existsMustBeDir(rulesFolder);
   const filesGrouped = groupFiles(rulesFolder);
 
   const rules = Object.entries(filesGrouped)
